% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\name{onet_crosswalk}
\alias{onet_crosswalk}
\title{Convert ONET SOC codes from any schema to any target schema}
\usage{
onet_crosswalk(
  df,
  onet_occupation_columnn,
  onet_soc_schema = "ONET_SOC_2019_CODE",
  keep_details = FALSE,
  include_soc = FALSE
)
}
\arguments{
\item{keep_details}{}
}
\description{
\code{onet_crosswalk} takes a df that contains a column of
ONET SOC codes and returns the same dataframe with additional columns for the
crosswalked SOC code.
}
