% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirl-functions.R
\name{pirl_name_replace}
\alias{pirl_name_replace}
\title{Replace PIRL element DF column names with friendly Names}
\usage{
pirl_name_replace(df)
}
\description{
\code{pirl_name_replace} takes a df with column names that are listed
as pirl elements and replaces them with friendly names
}
