% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\name{soc_crosswalk}
\alias{soc_crosswalk}
\title{Convert SOC codes from any schema to any target schema}
\usage{
soc_crosswalk(
  df,
  occupation_column,
  soc_schema = "OES_2019_Estimates_Code",
  keep_details = TRUE
)
}
\arguments{
\item{keep_details}{}
}
\description{
\code{soc_crosswalk} takes a df that contains a column of
SOC codes and returns the same dataframe with additional columns for the
crosswalked SOC code.
}
